fix: optimise index file
