+++ apps/web/components/Feature/Contents/ContentUploadModal/index.tsx
@@ -2,13 +2,8 @@
 
 import { useTranslation } from "react-i18next";
 import { BackButtonIcon } from "@/assets/icons";
-import SuccessArcIcon from "@/assets/icons/SuccessArcIcon";
-import UploadAudioIcon from "@/assets/icons/UploadAudioIcon";
-import UploadPdfIcon from "@/assets/icons/UploadPdfIcon";
-import UploadEpubIcon from "@/assets/icons/UploadEpubIcon";
 import GenericButton from "@/components/UI/GenericButton";
 import { GenericModal } from "@/components/UI/Modals";
-import GenericLoader from "@/components/UI/GenericLoader";
 import {
   HiddenInput,
   UploadHint,
@@ -16,7 +11,6 @@ import {
 } from "@/components/UI/ImageUploadCropModal/styles";
 import { BackButton } from "../ContentTypeModal/styles";
 import { VARIANT } from "@/utils/Constants";
-import { LOADER_SIZE, LOADER_VARIANT } from "@/utils/ui";
 import {
   ChooseUploadButton,
   ContentUploadDropZone,
@@ -24,17 +18,10 @@ import {
   UploadHelperText,
   UploadHelperTextGroup,
   UploadModalContent,
-  SelectedFileContainer,
-  FileDetailsWrapper,
-  PreviewFileRow,
-  FileInfoColumn,
-  PreviewBox,
-  PreviewVideo,
 } from "./styles";
 import { ContentType } from "@/utils/content";
 import { useContentUpload } from "@/hooks/contents/useContentUpload";
-import { formatFileSize } from "@/utils/file";
-import COLORS from "@repo/ui/colors";
+import SelectedFileView from "./SelectedFileView";
 
 type ContentUploadModalProps = {
   visible: boolean;
@@ -77,27 +64,6 @@ export default function ContentUploadModal({
     callback();
   };
 
-  const renderPreview = () => {
-    switch (uploadType) {
-      case "video":
-        return <PreviewVideo src={previewUrl ?? ""} controls={false} />;
-      case "audio":
-        return (
-          <UploadAudioIcon width={64} height={64} color={COLORS.primary.RED} />
-        );
-      case "pdf":
-        return (
-          <UploadPdfIcon width={64} height={64} color={COLORS.primary.RED} />
-        );
-      case "epub":
-        return (
-          <UploadEpubIcon width={64} height={64} color={COLORS.primary.BLUE} />
-        );
-      default:
-        return null;
-    }
-  };
-
   return (
     <GenericModal
       visible={visible}
@@ -121,37 +87,15 @@ export default function ContentUploadModal({
           />
 
           {selectedFile ? (
-            <SelectedFileContainer>
-              <UploadHint>Upload {uploadType}</UploadHint>
-
-              <FileDetailsWrapper>
-                <PreviewFileRow>
-                  <PreviewBox>{renderPreview()}</PreviewBox>
-
-                  <FileInfoColumn>
-                    <div>{selectedFile.name}</div>
-                    <div>{formatFileSize(selectedFile.size)}</div>
-                  </FileInfoColumn>
-                </PreviewFileRow>
-
-                {uploadComplete && <SuccessArcIcon width={32} height={32} />}
-
-                {isUploading && !uploadComplete && (
-                  <GenericLoader
-                    variant={LOADER_VARIANT.INLINE}
-                    size={LOADER_SIZE.SM}
-                  />
-                )}
-              </FileDetailsWrapper>
-
-              <GenericButton
-                variant={VARIANT.PRIMARY}
-                minWidth="320px"
-                disabled={!canProceed}
-              >
-                {t("common.next", { defaultValue: "Next" })}
-              </GenericButton>
-            </SelectedFileContainer>
+            <SelectedFileView
+              uploadType={uploadType}
+              selectedFile={selectedFile}
+              previewUrl={previewUrl}
+              isUploading={isUploading}
+              uploadComplete={uploadComplete}
+              canProceed={canProceed}
+              onNext={() => {}}
+            />
           ) : (
             <ContentUploadDropZone
               onDragOver={(e) => e.preventDefault()}
